// src/main/vue/contact-manager/src/store/request/mutation-types.ts
export enum MutationTypes {
  SET_REQUESTS = 'SET_REQUESTS',
  SET_STATE = 'SET_STATE',
  UPDATE_REQUEST = 'UPDATE_REQUEST',
  REMOVE_REQUEST = 'REMOVE_REQUEST',
  UPDATE_RECIPIENT = 'UPDATE_RECIPIENT',
  ADD_RECIPIENT = 'ADD_RECIPIENT',
  REMOVE_RECIPIENT = 'REMOVE_RECIPIENT',
  SET_TOTAL_COUNT = 'SET_TOTAL_COUNT',
  SET_MODAL_DATA = 'SET_MODAL_DATA',
}